New script to perform second pass over data
Second step goes through each position in each accession where
there was a TE insertion or deletion called in that region in
another accession, but not the inspected accession. Looks for split
reads linked to same TE as was identified as inserted or deleted in
other accessions. If a split read is found, TE insertions and deletions
files are updated with variant added.